Timezone fixes
Ok, some google calendars are coming back as UTC, some as user local
time. We only want to set them as UTC if they are not zone-aware, and
then convert to local time.

For all day events, we want them to start at midnight regardless of
timezone, so just set them to local time.